Decompose HomePage into focused hooks and components
Phase 3 of stabilization plan — HomePage reduced from 1,121 to 453 lines:

- Extract usePanelManager hook: single activePanel state replaces 4 booleans,
  eliminates bugs where opening one panel forgets to close another
- Extract useDocumentPersistence hook: save/load/auto-save/rename/delete logic
- Extract useRecentFiles hook: localStorage, pin/unpin, LRU, file limit logic
- Extract useEditorFormatting hook: insertFormatting, insertHeading, insertList,
  clearFormatting
- Extract RecentFilesSidebar component (132 lines of sidebar JSX)
- Extract FormattingToolbar component (172 lines of toolbar JSX)
- HomePage is now a composition root wiring hooks and components together
